pyproject.toml
[tool.poetry]
name = "auto-resume"
version = "0.1.0"
description = ""
authors = ["pandego <7780875+pandego@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ipykernel = "^6.29.5"
openai = "^1.43.0"
python-dotenv = "^1.0.1"
pre-commit = "^3.8.0"
jinja2 = "^3.1.4"
rich = "^13.8.0"
loguru = "^0.7.2"
langchain = "^0.2.16"
pdflatex = "^0.1.3"
pylatex = "^1.4.2"
pypdf2 = "^3.0.1"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
